backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pypdf>=4.2.0
Pillow>=10.3.0
pymupdf>=1.24.0
reportlab>=4.2.0
